requirements.txt
django==4.2
python-decouple==3.8
psycopg2==2.9.6
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0
celery==5.3.4
redis==5.0.1